% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGLM.R
\name{UTME}
\alias{UTME}
\title{Fit a Bayesian Uni-Trait Multi-Environment Regression Model}
\usage{
UTME(
  ETA = NULL,
  y = NULL,
  response_type = "DLN",
  nIter = 1000,
  nBurnin = 100,
  nThin = 1,
  R2 = 0.5,
  priorv = 5,
  priorS = NULL,
  priorvarbeta0 = NULL,
  rcontrol = 1.15,
  type_prediction = "mean",
  intercept = list(TRUE, 0),
  FIXED_r = TRUE,
  a0 = 0.01,
  c0 = 0.01,
  b0 = 0.01,
  g0 = 0.01,
  verbose = F,
  saveAt = ""
)
}
\arguments{
\item{ETA}{List used to specify the linear predictor.
By default it is set to NULL, in which case only the intercept is included.}

\item{y}{Response vector. Can contain NA values.}

\item{response_type}{Type of response ("DLN","gaussian","NB","Poisson","PLN").
For DLN, Poisson or PLN responses, the argument \code{y} must contain integer values,
while for gaussian responses any type of numeric value can be accepted.}

\item{nIter}{Number of MCMC iterations.}

\item{nBurnin}{Burn-in iterations.}

\item{nThin}{Thinning interval.}

\item{R2}{Proportion of variance explained a priori for the error term.}

\item{priorv}{Prior degrees of freedom for the error variance.}

\item{priorS}{Prior scale for the error variance.}

\item{priorvarbeta0}{Choose the variance, σ_0^2, to obtain a flat prior
for intercept,β_0∼N(0, σ_0^2). By default (internally) is 1e10.}

\item{rcontrol}{Specify the k-value used to control the difference between the
Conditional Mean and Conditional Variance of the Negative Binomial model.
Higher values of k indicate a smaller difference between the mean and variance,
approximating the equidispersion of a Poisson model.}

\item{type_prediction}{Used to specify if you want to predict with the posterior
mean or median in DLN model; default is mean.}

\item{intercept}{A list of length two. The first element is used to specify if
you want to include (TRUE) or not (FALSE) the intercept as an isolated term in
linear predictor; in case you specify FALSE, the second argument assigns the
position, according to the covariates order, where you want to include it.
The default values are list(TRUE,0).}

\item{FIXED_r}{Logical. If TRUE, a fixed Gamma prior is used for the NB dispersion
parameter,
r ~ Gamma(shape = a0, rate = c0).
If FALSE, a hierarchical prior is used,
r ~ Gamma(shape = a0, rate = h),
h ~ Gamma(shape = b0, rate = g0).}

\item{a0}{Shape parameter of the Gamma prior distribution for the NB dispersion
parameter r.}

\item{c0}{Rate parameter of the Gamma prior distribution for the NB dispersion
parameter r when \code{FIXED_r = TRUE}.}

\item{b0}{Shape parameter of the Gamma prior distribution for the hyperparameter h.}

\item{g0}{Rate parameter of the Gamma prior distribution for the hyperparameter h.}

\item{verbose}{Used to show (TRUE) or not (FALSE) the progress of iterations
in the Gibbs Sampler; default is FALSE}

\item{saveAt}{Used to indicate UTME where to store the samples and to provide
a pre-fix to be appended to the names of the file where samples are stored.
By default samples are saved in the current working directory and no pre-fix
is added to the file names}
}
\value{
A list containing posterior samples and summaries.
}
\description{
Fits a Bayesian Uni-trait and Multi-environment model using MCMC sampling.
Supports Gaussian, NB, Poisson, DLN and PLN responses.
}
