% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGM.R
\name{MTME}
\alias{MTME}
\title{Fit a Bayesian Multi-Trait Multi-Environment Regression Model}
\usage{
MTME(
  ETA = NULL,
  y = NULL,
  response_type = "DLN",
  nIter = 1000,
  nBurnin = 100,
  nThin = 1,
  R2 = 0.5,
  resCov = list(priorv = NULL, priorS = NULL, type = "UN"),
  priorSigma0 = NULL,
  rcontrol = 1.15,
  type_prediction = "mean",
  Iters_latent = 1,
  Burn_latent = 0,
  Thin_latent = 1,
  n_QMCpoints = 10,
  intercept_mt = T,
  verbose = F,
  saveAt = ""
)
}
\arguments{
\item{ETA}{List used to specify the linear predictor.
By default it is set to NULL, in which case only the intercept is included.}

\item{y}{Response vector. Can contain NA values.}

\item{response_type}{Type of response ("DLN","gaussian","Poisson","PLN").
For DLN, Poisson or PLN responses, the argument \code{y} must contain integer values,
while for gaussian responses any type of numeric value can be accepted.}

\item{nIter}{Number of MCMC iterations.}

\item{nBurnin}{Burn-in iterations.}

\item{nThin}{Thinning interval.}

\item{R2}{Proportion of variance explained a priori for the error term.}

\item{priorSigma0}{Choose the variance, Iσ_0^2, to obtain a flat prior
for the vector of intercepts,β_0∼MN(0, Iσ_0^2).By default (internally) is 1e10.}

\item{rcontrol}{Specify the k-value used to control the difference between the
Conditional Mean and Conditional Variance of the Negative Binomial model.
Higher values of k indicate a smaller difference between the mean and variance,
approximating the equidispersion of a Poisson model}

\item{type_prediction}{Used to specify if you want to predict with the posterior
mean or median in DLN model; default is mean.}

\item{Iters_latent}{Number of MCMC iterations to obtain Multivariate Truncated
normal samples.}

\item{Burn_latent}{Burn-in iterations for the Multivariate Truncated normal
samples.}

\item{Thin_latent}{Thinning interval for the Multivariate Truncated normal
samples.}

\item{n_QMCpoints}{Number of Sobol points for the Quasi-Monte Carlo method
used for the aproximation of the likelihood of the Multivariate DLN model.}

\item{verbose}{Used to show (TRUE) or not (FALSE) the progress of iterations
in the Gibbs Sampler; default is FALSE}

\item{saveAt}{Used to indicate UTME where to store the samples and to provide
a pre-fix to be appended to the names of the file where samples are stored.
By default samples are saved in the current working directory and no pre-fix
is added to the file names}

\item{priorv}{Prior degrees of freedom for the error co-variance matrix.}

\item{priorS}{Prior scale for the error co-variance matrix.}

\item{type}{Co-variance structure (Unstructured, Diagonal, Factor Analysis
or Recursive)}
}
\value{
A list containing posterior samples and summaries.
}
\description{
Fits a Bayesian Multi-trait and Multi-environment model using MCMC sampling.
Supports Gaussian, Poisson, DLN and PLN responses.
}
