% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGM.R
\name{UTME}
\alias{UTME}
\title{Fit a Bayesian Uni-Trait Multi-Environment Regression Model}
\usage{
UTME(
  ETA = NULL,
  y = NULL,
  response_type = "DLN",
  nIter = 1000,
  nBurnin = 100,
  nThin = 1,
  R2 = 0.5,
  priorv = 5,
  priorS = NULL,
  priorvarbeta0 = NULL,
  rcontrol = 1.15,
  type_prediction = "mean",
  intercept = list(TRUE, 0),
  verbose = F,
  saveAt = ""
)
}
\arguments{
\item{ETA}{List used to specify the linear predictor.
By default it is set to NULL, in which case only the intercept is included.}

\item{y}{Response vector. Can contain NA values.}

\item{response_type}{Type of response ("DLN","gaussian","Poisson","PLN").
For DLN, Poisson or PLN responses, the argument \code{y} must contain integer values,
while for gaussian responses any type of numeric value can be accepted.}

\item{nIter}{Number of MCMC iterations.}

\item{nBurnin}{Burn-in iterations.}

\item{nThin}{Thinning interval.}

\item{R2}{Proportion of variance explained a priori for the error term.}

\item{priorv}{Prior degrees of freedom for the error variance.}

\item{priorS}{Prior scale for the error variance.}

\item{priorvarbeta0}{Choose the variance, σ_0^2, to obtain a flat prior
for intercept,β_0∼N(0, σ_0^2).By default (internally) is 1e10.}

\item{rcontrol}{Specify the k-value used to control the difference between the
Conditional Mean and Conditional Variance of the Negative Binomial model.
Higher values of k indicate a smaller difference between the mean and variance,
approximating the equidispersion of a Poisson model}

\item{type_prediction}{Used to specify if you want to predict with the posterior
mean or median in DLN model; default is mean.}

\item{intercept}{A list of length two. The first element is a logical indicating
whether the intercept should be included. The second element indicates the
position of the intercept in the linear predictor.}

\item{verbose}{Used to show (TRUE) or not (FALSE) the progress of iterations
in the Gibbs Sampler; default is FALSE}

\item{saveAt}{Used to indicate UTME where to store the samples and to provide
a pre-fix to be appended to the names of the file where samples are stored.
By default samples are saved in the current working directory and no pre-fix
is added to the file names}
}
\value{
A list containing posterior samples and summaries.
}
\description{
Fits a Bayesian Uni-trait and Multi-environment model using MCMC sampling.
Supports Gaussian, Poisson, DLN and PLN responses.
}
